backend/requirements.txt
Flask==3.1.2
flask_cors==6.0.2
flask_sqlalchemy==3.1.1
python-dotenv==1.2.1
PyJWT==2.10.1
bcrypt==4.2.0
